% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deDuplicate.R
\name{deDuplicate}
\alias{deDuplicate}
\alias{deDuplicate.default}
\alias{deDuplicate.Scores}
\title{deDuplicate scored neighbors matrix.}
\usage{
deDuplicate(object, thresh, priority)

\method{deDuplicate}{default}(object, thresh, priority)

\method{deDuplicate}{Scores}(object, thresh, priority)
}
\arguments{
\item{object}{\code{\link{Scores}} object containing pre-computed Neighbors and match scores.}

\item{thresh}{The quantile at which to classify scores as duplicates.}

\item{priority}{A list of structure (Var= 'Varx', DESC=TRUE) with 'Varx'
being either a numeric or character value corresponding to a column in rawData.
This parameter determines how to sort the duplicates. 
DESC is a logical for this ordering (descending).}
}
\value{
An object of class \code{\link{Duplicates}} containing the scored neighbor pairs.
}
\description{
deDuplicate scored neighbors matrix.
}
