% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicates.R
\name{Duplicates}
\alias{Duplicates}
\title{Duplicates object.}
\usage{
Duplicates(ScoresObj, dupsList, details)
}
\arguments{
\item{ScoresObj}{\code{\link{Scores}} object.}

\item{dupsList}{A list with entries for each unique ID. This is a map
from the duplicate IDs to the unique IDs.}

\item{details}{Details to be printed regarding the thresholding and ordering}
}
\value{
An object of class \code{\link{Duplicates}}.
}
\description{
Stores the results from calling \code{\link{deDuplicate}}.
}
