% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedigrees}
\alias{pedigrees}
\title{Simulated family pedigrees.}
\format{An object of class \code{'cross'}; see \code{\link[qtl]{read.cross}}.}
\source{
\url{http://bcb.dfci.harvard.edu/bayesmendel/software.php/}
}
\usage{
data(pedigrees)
}
\description{
A dataset containing (possibly duplicated) family pedigrees, 
with each family consisting of a reported 'self' individual
and various cancer risk indicators/status, as well as a 
pseudo-random collection of reported family members
and their associated indicators. Allele sharing between generations
follow basic Mendelian inheritance with parameters provided by
BayesMendel.
}
\examples{
data(pedigrees)
self <- pedigrees[pedigrees$ID==1,]
\donttest{hist(self$nDuplicates,
   breaks=seq(0,max(sself$nDuplicates),by=1),
   main='Distribution of Duplications',xlab='Duplicate Count',
   ylab='Frequency',right=FALSE)}
}
\keyword{data}
\keyword{datasets}
\keyword{pedigrees}
