% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealTruth.R
\name{revealTruth}
\alias{revealTruth}
\alias{revealTruth.default}
\alias{revealTruth.list}
\title{Reveal scored neighbors matrix.}
\usage{
revealTruth(object, labels, duplabels)

\method{revealTruth}{default}(object, labels, duplabels)

\method{revealTruth}{list}(object, labels, duplabels)
}
\arguments{
\item{object}{\code{\link{Duplicates}} object to assess how the deduplication went.}

\item{labels}{character or numeric vector containing the true match labels with 
rows corresponding to the raw data matrix (in object[['rawData']])}

\item{duplabels}{character vector containing the duplicated labels 
(i.e. object[['ID']] from previous steps)}
}
\value{
An object of class \code{\link{Neighbors}} containing the scored neighbor pairs.
}
\description{
Reveal scored neighbors matrix.
}
