% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{Scores}
\alias{Scores}
\title{Scores object.}
\usage{
Scores(BlocksObj, Scores)
}
\arguments{
\item{BlocksObj}{\code{\link{Blocks}} object to score.}

\item{Scores}{Numeric vector of match scores.}
}
\value{
An object of class \code{\link{Scores}}.
}
\description{
Stores the scored neighbors and parameters from calling 
\code{\link{calculateScore}}.
}
