% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealTruth.R
\name{revealTruth}
\alias{revealTruth}
\alias{revealTruth.default}
\alias{revealTruth.Duplicates}
\title{Reveal scored neighbors matrix.}
\usage{
revealTruth(object, truelabels, duplabels)

\method{revealTruth}{default}(object, truelabels, duplabels)

\method{revealTruth}{Duplicates}(object, truelabels, duplabels)
}
\arguments{
\item{object}{object of class \code{\link{Duplicates}} to assess how the deduplication went.}

\item{truelabels}{character or numeric vector containing the true match labels with
rows corresponding to the raw data matrix (in object[['rawData']])}

\item{duplabels}{character vector containing the duplicated labels
(i.e. object[['ID']] from previous steps)}
}
\value{
An object of class \code{\link{Neighbors}} containing the scored neighbor pairs.
}
\description{
Reveal scored neighbors matrix.
}
